chore(mariadb-monitor): Cleanup and structure the code
