+++ libs/mariadb_monitor/metrics.go
@@ -4,10 +4,10 @@ import (
 	"bufio"
 	"context"
 	"fmt"
+	"log/slog"
 	"net"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
@@ -21,6 +21,21 @@ type MemoryStatus struct {
 	SwapFree     uint64
 	UsagePercent float64
 	SwapPercent  float64
+}
+
+type cpuTimes struct {
+	user, nice, system, idle, iowait, irq, softirq, steal uint64
+}
+
+func (c cpuTimes) total() uint64 {
+	return c.user + c.nice + c.system + c.idle + c.iowait + c.irq + c.softirq + c.steal
+}
+
+type snapshotCache struct {
+	prevCPU       *cpuTimes
+	prevPgIn      uint64
+	prevPgOut     uint64
+	hasPgSnapshot bool
 }
 
 func checkPSI(resource string) (float64, error) {
@@ -54,14 +69,6 @@ func checkPSI(resource string) (float64, error) {
 	return 0, fmt.Errorf("no 'some' line found in %s", path)
 }
 
-type cpuTimes struct {
-	user, nice, system, idle, iowait, irq, softirq, steal uint64
-}
-
-func (c cpuTimes) total() uint64 {
-	return c.user + c.nice + c.system + c.idle + c.iowait + c.irq + c.softirq + c.steal
-}
-
 func readCPUTimes() (cpuTimes, error) {
 	f, err := os.Open("/proc/stat")
 	if err != nil {
@@ -91,13 +98,6 @@ func readCPUTimes() (cpuTimes, error) {
 		return ct, nil
 	}
 	return cpuTimes{}, fmt.Errorf("no 'cpu' line found in /proc/stat")
-}
-
-type snapshotCache struct {
-	prevCPU       *cpuTimes
-	prevPgIn      uint64
-	prevPgOut     uint64
-	hasPgSnapshot bool
 }
 
 func checkIOWait(cache *snapshotCache) (float64, error) {
@@ -232,12 +232,12 @@ func checkIOFreeze(timeout time.Duration) bool {
 	done := make(chan bool, 1)
 
 	go func() {
-		path := filepath.Join(os.TempDir(), ".mariadb_monitor_io_test")
-		f, err := os.Create(path)
+		f, err := os.CreateTemp("", "mariadb_monitor_io_test_*")
 		if err != nil {
 			done <- false
 			return
 		}
+		path := f.Name()
 		_, err = f.Write([]byte("io_test"))
 		if err != nil {
 			f.Close()
@@ -248,14 +248,14 @@ func checkIOFreeze(timeout time.Duration) bool {
 		_ = f.Sync()
 		f.Close()
 		os.Remove(path)
-		done <- false // I/O completed, not frozen
+		done <- false
 	}()
 
 	select {
 	case frozen := <-done:
 		return frozen
 	case <-time.After(timeout):
-		return true // timed out, I/O is frozen
+		return true
 	}
 }
 
@@ -265,17 +265,14 @@ func checkMachineFrozen() (bool, string) {
 
 	deadline, ok := ctx.Deadline()
 	if !ok {
-		// Fallback: context should always have a deadline here, but be defensive.
 		deadline = time.Now().Add(2 * time.Second)
 	}
 
-	// 1. Check process spawning (fork/exec). Tests PID limits, memory for fork, and basic scheduling.
 	cmd := exec.CommandContext(ctx, "true")
 	if err := cmd.Run(); err != nil {
 		return true, "fork_failed"
 	}
 
-	// 2. Check File Descriptor (FD) limits. SSH often hangs when the OS runs out of file handles.
 	r, w, err := os.Pipe()
 	if err != nil {
 		return true, "fd_exhausted"
@@ -283,15 +280,12 @@ func checkMachineFrozen() (bool, string) {
 	r.Close()
 	w.Close()
 
-	// 3. Check Network Stack and Memory limits. 
-	// Verifies if the OS has enough memory to open, dial, and accept TCP connections.
 	ln, err := net.Listen("tcp", "127.0.0.1:0")
 	if err != nil {
 		return true, "tcp_listen_failed"
 	}
 	defer ln.Close()
 
-	// Ensure we still have time left in the overall budget before network checks.
 	remaining := time.Until(deadline)
 	if remaining <= 0 {
 		return true, "deadline_exceeded"
@@ -306,7 +300,6 @@ func checkMachineFrozen() (bool, string) {
 		errCh <- err
 	}()
 
-	// Dial timeout is derived from the remaining overall budget, capped at 1s.
 	dialTimeout := remaining
 	if dialTimeout > time.Second {
 		dialTimeout = time.Second
@@ -318,7 +311,6 @@ func checkMachineFrozen() (bool, string) {
 	}
 	conn.Close()
 
-	// Recompute remaining time for the accept phase, again capped at 1s.
 	remaining = time.Until(deadline)
 	if remaining <= 0 {
 		return true, "deadline_exceeded"
@@ -338,4 +330,173 @@ func checkMachineFrozen() (bool, string) {
 	}
 
 	return false, ""
+}
+
+type metricWindows struct {
+	psiCPU       MetricWindow
+	psiMemory    MetricWindow
+	psiIO        MetricWindow
+	iowait       MetricWindow
+	memUsage     MetricWindow
+	swapUsage    MetricWindow
+	pageRate     MetricWindow
+	memAvailable MetricWindow
+	memCached    MetricWindow
+	swapFree     MetricWindow
+}
+
+func newMetricWindows(size int) *metricWindows {
+	return &metricWindows{
+		psiCPU:       NewMetricWindow(size),
+		psiMemory:    NewMetricWindow(size),
+		psiIO:        NewMetricWindow(size),
+		iowait:       NewMetricWindow(size),
+		memUsage:     NewMetricWindow(size),
+		swapUsage:    NewMetricWindow(size),
+		pageRate:     NewMetricWindow(size),
+		memAvailable: NewMetricWindow(size),
+		memCached:    NewMetricWindow(size),
+		swapFree:     NewMetricWindow(size),
+	}
+}
+
+type frozenState struct {
+	frozen bool
+	reason string
+}
+
+func (m *Monitor) collectTriggers() ([]string, *frozenState) {
+	var triggers []string
+	var frozen *frozenState
+
+	if val, err := checkPSI("cpu"); err == nil {
+		m.windows.psiCPU.Push(val, m.cfg.PSICPUThreshold)
+		slog.Debug("psi_cpu", "avg10", val)
+	} else {
+		slog.Warn("failed to check psi cpu", "error", err)
+	}
+
+	if val, err := checkPSI("memory"); err == nil {
+		m.windows.psiMemory.Push(val, m.cfg.PSIMemoryThreshold)
+		slog.Debug("psi_memory", "avg10", val)
+	} else {
+		slog.Warn("failed to check psi memory", "error", err)
+	}
+
+	if val, err := checkPSI("io"); err == nil {
+		m.windows.psiIO.Push(val, m.cfg.PSIIOThreshold)
+		slog.Debug("psi_io", "avg10", val)
+	} else {
+		slog.Warn("failed to check psi io", "error", err)
+	}
+
+	iowaitVal, iowaitErr := checkIOWait(m.cache)
+	if iowaitErr == nil {
+		m.windows.iowait.Push(iowaitVal, m.cfg.IOWaitThreshold)
+		slog.Debug("iowait", "percent", fmt.Sprintf("%.1f", iowaitVal))
+	} else {
+		slog.Warn("failed to check iowait", "error", iowaitErr)
+	}
+
+	mem, memErr := checkMemory()
+	if memErr == nil {
+		m.windows.memUsage.Push(mem.UsagePercent, m.cfg.MemoryUsageThreshold)
+		m.windows.swapUsage.Push(mem.SwapPercent, m.cfg.SwapUsageThreshold)
+		m.windows.memAvailable.Push(float64(mem.MemAvailable), 0)
+		m.windows.memCached.Push(float64(mem.Cached), 0)
+		m.windows.swapFree.Push(float64(mem.SwapFree), 0)
+
+		slog.Debug("memory",
+			"usage_percent", fmt.Sprintf("%.1f", mem.UsagePercent),
+			"swap_percent", fmt.Sprintf("%.1f", mem.SwapPercent),
+			"mem_available_kb", mem.MemAvailable,
+			"cached_kb", mem.Cached,
+			"swap_free_kb", mem.SwapFree,
+		)
+	} else {
+		slog.Warn("failed to check memory", "error", memErr)
+	}
+
+	pageRateVal, pageRateErr := checkPageRate(m.cache, m.cfg.CheckInterval)
+	if pageRateErr == nil {
+		m.windows.pageRate.Push(pageRateVal, m.cfg.PageRateThreshold)
+		slog.Debug("page_rate", "pages_per_sec", pageRateVal)
+	} else {
+		slog.Warn("failed to check page rate", "error", pageRateErr)
+	}
+
+	if memErr == nil && mem.UsagePercent >= m.cfg.CriticalMemoryThreshold {
+		triggers = append(triggers, fmt.Sprintf("critical_memory=%.1f%%", mem.UsagePercent))
+	}
+
+	if iowaitErr == nil && iowaitVal >= m.cfg.IOWaitThreshold {
+		if checkIOFreeze(m.cfg.IOFreezeTimeout) {
+			triggers = append(triggers, "io_freeze")
+		}
+	}
+
+	if m.windows.psiCPU.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_psi_cpu")
+	}
+	if m.windows.psiMemory.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_psi_memory")
+	}
+	if m.windows.psiIO.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_psi_io")
+	}
+	if m.windows.iowait.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_iowait")
+	}
+	if m.windows.memUsage.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_memory_usage")
+	}
+	if m.windows.swapUsage.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_swap_usage")
+	}
+	if m.windows.pageRate.IsSustained(m.cfg.SustainedRatio) {
+		triggers = append(triggers, "sustained_page_rate")
+	}
+
+	if isFrozen, reason := checkMachineFrozen(); isFrozen {
+		frozen = &frozenState{frozen: true, reason: reason}
+		triggers = append(triggers, fmt.Sprintf("machine_frozen(%s)", reason))
+	}
+
+	if memErr == nil && mem.SwapTotal > 0 {
+		swapFreePercent := float64(mem.SwapFree) / float64(mem.SwapTotal) * 100.0
+		if m.windows.memAvailable.Trend() == "falling" &&
+			m.windows.memCached.Trend() == "falling" &&
+			swapFreePercent < m.cfg.SwapHeadroom {
+			triggers = append(triggers, fmt.Sprintf("predictive_memory_exhaustion(swap_free=%.1f%%)", swapFreePercent))
+		}
+	}
+
+	logSpikes(m.cfg, m.windows)
+	return triggers, frozen
+}
+
+func logSpikes(cfg Config, w *metricWindows) {
+	type spike struct {
+		name   string
+		window *MetricWindow
+	}
+	spikes := []spike{
+		{"psi_cpu", &w.psiCPU},
+		{"psi_memory", &w.psiMemory},
+		{"psi_io", &w.psiIO},
+		{"iowait", &w.iowait},
+		{"memory_usage", &w.memUsage},
+		{"swap_usage", &w.swapUsage},
+		{"page_rate", &w.pageRate},
+	}
+	for _, s := range spikes {
+		if s.window.IsSpike(cfg.SustainedRatio) {
+			latest := s.window.buf.Latest()
+			slog.Info("spike detected",
+				"metric", s.name,
+				"value", fmt.Sprintf("%.1f", latest.value),
+				"trend", s.window.Trend(),
+			)
+		}
+	}
 }
